requirements.txt
# AI Coach POC Requirements

# Core Dependencies
anthropic>=0.25.0
openai>=1.12.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Async Support
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Environment Management
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0

# Scientific Computing
scipy>=1.10.0

# Visualization (for metrics analysis)
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional Performance Enhancements
orjson>=3.8.0  # Faster JSON parsing
uvloop>=0.17.0; sys_platform != "win32"  # Faster event loop (Unix/Mac only)

# Error Tracking (Optional)
sentry-sdk>=1.40.0